Simplify code calculating parents and children
I think the sorting we did previously was to handle spans that were out of order, either by being appended or by being
discontinuous. We've recently removed the ability to append, and we haven't implemented discontinuous spans yet, so for
now it should be fine to remove this extra sorting.